@@ -11,7 +11,8 @@ namespace Assets.Scripts.MapEditor
         None = 0,
         Walkable = 1,
         Water = 2,
-        Snipable = 4
+        Snipable = 4,
+        SeeThrough = Walkable | Snipable,
     }
 
     [Serializable]
@@ -33,6 +34,7 @@ namespace Assets.Scripts.MapEditor
         public RoWalkCell Cell(int x, int y) => Cells[x + y * Width];
         public RoWalkCell Cell(Vector2Int pos) => Cells[pos.x + pos.y * Width];
         public bool CellWalkable(int x, int y) => (Cell(x, y).Type & CellType.Walkable) == CellType.Walkable;
+        public bool DoesCellBlockLos(int x, int y) => (Cell(x, y).Type & CellType.SeeThrough) == 0;
         
 
         public static CellType ColorToCellMask(string color)
