@@ -6,7 +6,7 @@ namespace Assets.Scripts.SkillHandlers
 	{
 		static ClientSkillHandler()
 		{
-			handlers = new SkillHandlerBase[134];
+			handlers = new SkillHandlerBase[135];
 			handlers[0] = new DefaultSkillHandler();
 			handlers[1] = new DefaultSkillHandler();
 			handlers[2] = new FirstAidHandler();
@@ -89,26 +89,26 @@ namespace Assets.Scripts.SkillHandlers
 			handlers[72] = new LordOfVermilionHandler();
 			handlers[73] = new DefaultSkillHandler();
 			handlers[74] = new DefaultSkillHandler();
-			handlers[75] = new FireAttackHandler();
-			handlers[76] = new IceAttackHandler();
-			handlers[77] = new WaterAttackHandler();
-			handlers[78] = new WindAttackHandler();
-			handlers[79] = new EarthAttackHandler();
-			handlers[80] = new DarkAttackHandler();
-			handlers[81] = new HolyAttackHandler();
-			handlers[82] = new PoisonAttackHandler();
-			handlers[83] = new UndeadAttackHandler();
-			handlers[84] = new GhostAttackHandler();
-			handlers[85] = new CriticalSlashHandler();
-			handlers[85].ExecuteWithoutSource = true;
-			handlers[86] = new DefaultSkillHandler();
-			handlers[87] = new PoisonHandler();
-			handlers[88] = new SleepAttackHandler();
-			handlers[89] = new StunAttackHandler();
-			handlers[89].ExecuteWithoutSource = true;
-			handlers[90] = new DefaultSkillHandler();
-			handlers[91] = new SilenceAttackHandler();
-			handlers[92] = new DefaultSkillHandler();
+			handlers[75] = new DefaultSkillHandler();
+			handlers[76] = new FireAttackHandler();
+			handlers[77] = new IceAttackHandler();
+			handlers[78] = new WaterAttackHandler();
+			handlers[79] = new WindAttackHandler();
+			handlers[80] = new EarthAttackHandler();
+			handlers[81] = new DarkAttackHandler();
+			handlers[82] = new HolyAttackHandler();
+			handlers[83] = new PoisonAttackHandler();
+			handlers[84] = new UndeadAttackHandler();
+			handlers[85] = new GhostAttackHandler();
+			handlers[86] = new CriticalSlashHandler();
+			handlers[86].ExecuteWithoutSource = true;
+			handlers[87] = new DefaultSkillHandler();
+			handlers[88] = new PoisonHandler();
+			handlers[89] = new SleepAttackHandler();
+			handlers[90] = new StunAttackHandler();
+			handlers[90].ExecuteWithoutSource = true;
+			handlers[91] = new DefaultSkillHandler();
+			handlers[92] = new SilenceAttackHandler();
 			handlers[93] = new DefaultSkillHandler();
 			handlers[94] = new DefaultSkillHandler();
 			handlers[95] = new DefaultSkillHandler();
@@ -118,39 +118,40 @@ namespace Assets.Scripts.SkillHandlers
 			handlers[99] = new DefaultSkillHandler();
 			handlers[100] = new DefaultSkillHandler();
 			handlers[101] = new DefaultSkillHandler();
-			handlers[102] = new SelfDestructHandler();
-			handlers[102].ExecuteWithoutSource = true;
-			handlers[103] = new PowerUpHandler();
-			handlers[104] = new SpeedUpHandler();
-			handlers[105] = new CallMinionHandler();
-			handlers[106] = new DefaultSkillHandler();
+			handlers[102] = new DefaultSkillHandler();
+			handlers[103] = new SelfDestructHandler();
+			handlers[103].ExecuteWithoutSource = true;
+			handlers[104] = new PowerUpHandler();
+			handlers[105] = new SpeedUpHandler();
+			handlers[106] = new CallMinionHandler();
 			handlers[107] = new DefaultSkillHandler();
 			handlers[108] = new DefaultSkillHandler();
-			handlers[109] = new BloodDrainHandler();
-			handlers[110] = new EnergyDrainHandler();
-			handlers[111] = new GuidedAttackHandler();
-			handlers[112] = new DefaultSkillHandler();
-			handlers[113] = new ComboAttackHandler();
-			handlers[114] = new DefaultSkillHandler();
-			handlers[115] = new DarkStrikeHandler();
-			handlers[116] = new DefaultSkillHandler();
+			handlers[109] = new DefaultSkillHandler();
+			handlers[110] = new BloodDrainHandler();
+			handlers[111] = new EnergyDrainHandler();
+			handlers[112] = new GuidedAttackHandler();
+			handlers[113] = new DefaultSkillHandler();
+			handlers[114] = new ComboAttackHandler();
+			handlers[115] = new DefaultSkillHandler();
+			handlers[116] = new DarkStrikeHandler();
 			handlers[117] = new DefaultSkillHandler();
 			handlers[118] = new DefaultSkillHandler();
 			handlers[119] = new DefaultSkillHandler();
-			handlers[120] = new SmokingHandler();
-			handlers[121] = new DefaultSkillHandler();
+			handlers[120] = new DefaultSkillHandler();
+			handlers[121] = new SmokingHandler();
 			handlers[122] = new DefaultSkillHandler();
 			handlers[123] = new DefaultSkillHandler();
 			handlers[124] = new DefaultSkillHandler();
 			handlers[125] = new DefaultSkillHandler();
-			handlers[126] = new BulwarkHandler();
-			handlers[127] = new DefaultSkillHandler();
-			handlers[128] = new GrandThunderstormHandler();
-			handlers[129] = new DefaultSkillHandler();
+			handlers[126] = new DefaultSkillHandler();
+			handlers[127] = new BulwarkHandler();
+			handlers[128] = new DefaultSkillHandler();
+			handlers[129] = new GrandThunderstormHandler();
 			handlers[130] = new DefaultSkillHandler();
 			handlers[131] = new DefaultSkillHandler();
 			handlers[132] = new DefaultSkillHandler();
 			handlers[133] = new DefaultSkillHandler();
+			handlers[134] = new DefaultSkillHandler();
 		}
 	}
 }
